detect bad replies from an enqueue operation and retry
When a fiber elsewhere within the same thread exits, the last
fiber to call #transfer is where execution restarts. This lead to
many occurrences where an enqueue operation was waiting for a valid
reply but instead got a nil. To fix, we detect the nil and ask the
scheduler to reschedule this fiber. When a valid reply comes along
it will be delivered back to here. Note that this rescheduling may
occur multiple times if lots of fibers are exiting.

This fixed many crashes!