found a strange bug in MRI 2.5.0 where closures and fibers have a weird interaction. I noticed in the examples/echo_tcp.rb runs that sometimes an echo message going back from server to client would get lost. Looking into it I confirmed that the server-side socket was getting closed, so the client was returning an empty string.
Looking deeper showed that the ensure block on TCP#accept was
running before it should. When the contents of the accept block
would suspend on a timer sleep, another fiber was picking up
execution of that same block after the yield. This made no sense.
Diving deeper and instrumenting everything proved that this was
not an issue of a fiber or block exiting and then another fiber
resuming from a strange place. Instead, it appeared that the
block closure was getting confused. I know this makes little
sense. No easy way to repro with a small code example but this
behavior DID occur regularly (80% of the runs).

The fix was simple. Much like I discovered some weirdness in the
Fiber Scheduler when creating a block closure, I decided to move
the creation of the encompassing block to its own method. The
specific variables I wanted the closure to capture were then
passed as arguments. This made it "clean."

My guess is that the block in #accept was getting confused by
the socket, address, rc, and errno variables which were all
shadowed by vars of the same name outside of a block. The closure
was maybe grabbing references to the outer vars.