define EPollEventStruct as a packed struct to fix major bug
Struct should be 12 bytes but it was allocated as 16. By default
FFI aligns to 8-byte boundaries, so the struct was too large.
This was causing epoll_ctl and epoll_wait to set bits in the wrong
place, so garbage in gets you garbage out. Now struct aligns fields
on 1-byte boundaries.

Second, the bitwise flag comparisons now use "!= 0" instead of
"> 0". Minor, but this is more correct.

Third, reduced the number of fields in the EPollDataUnion to two.
I really only use one, so the second one is only necessary to make
sure the struct is padded to the right size.