begin transitioning to new POSIX namespace
All platform-specific functions, constants, and structs will be
created in IO::Platforms::Functions, IO::Platforms::Constants,
and IO::Platforms::Structs respectively. Then this source
information will be used in the flat IO::POSIX namespace to
contain everything for easy reference.

Items that remain platform-specific such as #kqueue and #epoll
will remain in the IO::Platforms namespace.

Further, some platforms define structs slightly differently even
though they are POSIX compliant. For example, the sockaddr_in
struct contains a :sin_len field on BSD UNIX but that field does
not exist on Linux. Likewise, sometimes fields are rearranged in
the header files for BSD vs Linux, so the FFI::Struct layout
needs to be defined per platform.

Ultimately a lot of this will be script generated for the
platforms, but for now it makes sense to rearrange the namespace
for clarity.